GPU: add an emulated binary64 and a selector for GPUdoubleCalc
GPUdoubleCalc now names one of four representations, chosen by GPUCA_DOUBLECALC:
hardware double, plain float, the compensated two-float type, or a full IEEE-754
binary64 in software. The default is unchanged, double everywhere except Metal
and the two-float type there, so this is a no-op for every existing build. It
replaces GPUCA_FORCE_DOUBLECALC and GPUCA_FORCE_FLOATCALC, which could only
express two of the four.

The binary64 emulation is round to nearest even with subnormals, infinities and
NaNs, and correctly rounded division; NaN propagation follows the ARM64 order so
an Apple host is a reference down to the payload. It is not for production: over
the whole propagateTo kernel on an M1 Max it costs of the order of a hundred
times plain float, against roughly two for the two-float type. It is here because it
reproduces the CPU result exactly -- every covariance element bit-identical --
which turns a disagreement between Metal and the CPU into a search over code
rather than over numerics.

Measured on that kernel, 524288 tracks with strongly correlated covariances,
median of ten runs, error as |dC_ij| / sqrt(C_ii C_jj) against the CPU double
result: plain float 0.108 ns and 1.2e-5, two-float 0.256 ns and 5.6e-7, binary64
12.2 ns and zero. Only the binary64 kernel is unstable run to run, between 9.1
and 12.3 ns; the others vary by a few percent. For scale, the CPU double result
is itself 1.1e-6 from a cancellation-free reference, so the two-float type is
already at the noise floor of the formula.

For scale in the other direction, the same propagateTo on this machine's CPU,
where double is native, costs 23.5 ns in float and 27.1 ns in double, so real
hardware double is a 1.15x proposition. None of this arithmetic would be needed
if Metal had it.

Validated against hardware double on 9M operand pairs per operation, over uniform
bit patterns, near-equal exponents, the subnormal band, sparse mantissas for the
exact and halfway cases, the special values and track-like magnitudes, plus 4M
conversions each way: no mismatches.

The arithmetic must stay out of line on Metal. Inlined into a kernel it drops the
occupancy and runs two to five times slower than the call.

+++ DataFormats/Reconstruction/src/TrackParametrizationWithError.cxx
@@ -1087,7 +1087,7 @@ GPUd() auto TrackParametrizationWithError<value_T>::getPredictedChi2(const value
   auto chi2 = (d * (szz * d - sdz * z) + z * (sdd * z - d * sdz)) / det;
   if (chi2 < 0.) {
 #ifndef GPUCA_ALIGPUCODE
-    LOGP(warning, "Negative chi2={}, Cluster: {} {} {} Dy:{} Dz:{} | sdd:{} sdz:{} szz:{} det:{}", chi2, cov[0], cov[1], cov[2], d, z, sdd, sdz, szz, det);
+    LOGP(warning, "Negative chi2={}, Cluster: {} {} {} Dy:{} Dz:{} | sdd:{} sdz:{} szz:{} det:{}", double(chi2), cov[0], cov[1], cov[2], d, z, double(sdd), double(sdz), double(szz), double(det));
     LOGP(warning, "Track: {}", asString());
 #endif
   }
